Always add the manage.py directory to PYTHONPATH
Checking for the manage.py directory being the current directory is not
actually needed. If the path ends up being included twice it isn't a very big
issue (the first match still wins), but if you want to store notebooks in a
directory other than the root of the mange.py directory you currently have to
use relative imports to achieve this. This change fixes that so you can
always refer to your django package as if it was an installed package.

Fixes: #865