Correct grammar for django-extensions
There were some grammar issues with the documentation:
-----BEFORE-----

Django 2.0 now contains similar functionality as ForeignKeyAutocompleteAdmin therefor we are deprecated this extension and high encourage everybody to update to it.
This code will be removed in the near feature when support for Django older then 2.0 is dropped.
-----AFTER---------
Django 2.0 now contains similar functionality as *ForeignKeyAutocompleteAdmin* therefore we are deprecating this extension and highly encouraging everyone to update to it.

----